discord/role_assign/requirements.txt
click~=8.1.3
discord~=2.2.2
discord.py @ git+https://github.com/Rapptz/discord.py@dc9fb1fd84914df64576264d9cf6c151e768fefb
python-decouple==3.7
SQLAlchemy~=2.0.17
pymysql==1.0.3
pytz~=2023.3
requests~=2.31.0
Pillow==9.4.0
numerize==0.12
better-profanity==0.7.0
aiohttp~=3.7.4.post0
pandas
